Replace filePath parameter with data in hashObject
The hashObject function now accepts data as its parameter rather than
a file path.
This change is necessary because, when creating a tree object, there is
no singular file path(a single file) that can be passed as a parameter to
the hashObject function.
In case of blobs, the file path of the file being hashed can be passed.
But for tree objects, there is no such singular file.