Implement read-tree subcommand
The read-tree subcommand is used to recreate a tree(directory) from an objeect
id(tree object-id).
The parseTreeObject function  takes in a tree object and splits the entries in the
tree object into a object type, object id, and object name.
The readTree function takes in an object id(tree object id) as a parameter, internally
uses the parseTreeObject and recreates the tree whose object id was passed as an
argument.