Add commit subcommand
The commit subcommand takes a snapshot of the directory(using write-tree),
and creates a commit object and stores in the object database.
The commit object consists of key-value pairs(currently only tree followed
by the object id of snapshot) and a commit message separated from the
key-value paris by a new line.