Implement getOid function
The getOid function can be used to obtain the object-id associated
with a tag.
Its flexible, a user can provide either <tag>, tags/<tag>, or
ref/tags/<tag> to fetch their object.
If the function is given an object-id as a parameter, it returns
the same object-id, provided its a valid object-id.