Fix bug in tag subcommand
The name argument was set to "@" by default. The intention
behind this was to tag the HEAD commit by default, and as @ is an
alias for HEAD, name was set to @.
But this is incorrect as name refers to the name of the tag and not
the commit which is to be tagged.