Generalize iterParentsAndCommits function
The iterParentsAndCommits function has been generalized to now be used
with the log subcommand.
Since the log subcommand requires the oids returned to be in a deterministic
order(an oid must return the oids of its parents in sequence), a deque is now used to
store the oids, thereby maintaing the sequence of relations.