Create RefValue container to represent references
The references are now represented using the RefValue container.
The symbolic field indicates whether the reference is symbolic or not.
The value is the value stored by the reference.
Additionally, the getObject function previously didn't convert the
objectId parameter to an oid(the objectId parameter maybe a reference too),
there was a problem in visualising the references using the k sub-command.
Fixed this issue.