Bug Fix - Resolve start_point in branch
The start_point needs to be resolved to an oid before creating
a branch.
If this isn't done, "@" or any other ref will be written to the branch
file, which can lead to inconsistencies.
E.g.: If "@" is written to the branch file, at commit 'x' and head moves
forward, the branch file is still pointing to head, which points to a new
commit rather than commit 'x'.