Bug Fix - Return reference passed if deref is false
If the deref has been set to false, we would want _getRefInternal
to return the reference we passed along with RefValue.
Previously, the value of the reference alond with RefValue was being
returned even when deref was set to False.
This leads to problems, such as:
When updating a ref, the updateRef function uses getRef, which uses
_getRefInternal. As a consequence, even when deref is False, the reference
being pointed to ends up being updated rather than the original reference.