Create new reference file if reference doesn't exist
Added code to create a reference file if the reference doesn't
exist already.
Since _getRefInternal requires a reference to exist, a existing
reference must be passed to it. Therefore, its necessary to make
a reference, if it doesn't exist, before calling _getRefInternal.
If deref is set to true, the ref path needs to be updated to the
pointed file.