Implement readTreeMerge-Partial branch merge
The merge sub-command can now be used to merge commits from
two different branches onto a branch.
The command invokes base.merge which invokes base.readTreeMerge.
This functions creates the blobs for the files present in both
the commits, and writes these files to the current working directory.
It doesn't create a different commit as of now. This part of the merge
functionality is still pending.